requirements.txt
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
langchain-huggingface==0.0.3
langchain-core==0.2.0
python-dotenv==1.0.1
unidecode==1.3.8
chromadb==0.5.0
pypdf==4.2.0
streamlit==1.35.0
duckduckgo-search==6.1.0
sentence-transformers==3.0.0
openai==1.30.0
tiktoken==0.7.0
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
